Update appearance fetching to match new GraphQL schema
The appearance fields are fetching using a fragment to allow more granular type generation, as explained here: https://github.com/dotansimha/graphql-code-generator/issues/1602#issuecomment-478315105

In short, we want to import a type containing exactly the appearance fields that are fetching in the ContestQuery, and use that type for the PerformanceRow input. Perhaps the code generator will solve this better in the future, but for now fragments are an ok solution.